+++ src/Doctrine/Orm/RouteProvider.php
@@ -19,6 +19,7 @@ use Symfony\Cmf\Component\Routing\Candidates\CandidatesInterface;
 use Symfony\Cmf\Component\Routing\RouteProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Exception\RouteNotFoundException;
+use Symfony\Component\Routing\Route as SymfonyRoute;
 use Symfony\Component\Routing\RouteCollection;
 
 /**
@@ -32,10 +33,7 @@ use Symfony\Component\Routing\RouteCollection;
  */
 class RouteProvider extends DoctrineProvider implements RouteProviderInterface
 {
-    /**
-     * @var CandidatesInterface
-     */
-    private $candidatesStrategy;
+    private CandidatesInterface $candidatesStrategy;
 
     public function __construct(ManagerRegistry $managerRegistry, CandidatesInterface $candidatesStrategy, $className)
     {
@@ -46,7 +44,7 @@ class RouteProvider extends DoctrineProvider implements RouteProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getRouteCollectionForRequest(Request $request)
+    public function getRouteCollectionForRequest(Request $request): RouteCollection
     {
         $collection = new RouteCollection();
 
@@ -66,7 +64,7 @@ class RouteProvider extends DoctrineProvider implements RouteProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getRouteByName($name)
+    public function getRouteByName($name): SymfonyRoute
     {
         if (!$this->candidatesStrategy->isCandidate($name)) {
             throw new RouteNotFoundException(sprintf('Route "%s" is not handled by this route provider', $name));
@@ -83,7 +81,7 @@ class RouteProvider extends DoctrineProvider implements RouteProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getRoutesByNames($names = null)
+    public function getRoutesByNames($names = null): array
     {
         if (null === $names) {
             if (0 === $this->routeCollectionLimit) {
@@ -110,10 +108,7 @@ class RouteProvider extends DoctrineProvider implements RouteProviderInterface
         return $routes;
     }
 
-    /**
-     * @return ObjectRepository
-     */
-    protected function getRouteRepository()
+    protected function getRouteRepository(): ObjectRepository
     {
         return $this->getObjectManager()->getRepository($this->className);
     }