+++ src/Doctrine/Phpcr/PrefixInterface.php
@@ -20,13 +20,12 @@ namespace Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr;
 interface PrefixInterface
 {
     /**
-     * @return string the full path of this document in the repository
+     * Return the full path of this document in the repository.
      */
-    public function getId();
+    public function getId(): string;
 
     /**
-     * @param string $prefix The path in the repository to the routing root
-     *                       document
+     * Set the path in the repository to the routing root document.
      */
-    public function setPrefix($prefix);
+    public function setPrefix(string $prefix): static;
 }